Split out Panel Container into its own file
