+++ components/brave_wallet_ui/panel/wallet_panel.tsx
@@ -5,38 +5,14 @@
 
 import * as React from 'react'
 import { render } from 'react-dom'
-import { connect, Provider } from 'react-redux'
-import { bindActionCreators, Dispatch } from 'redux'
+import { Provider } from 'react-redux'
 
 import walletDarkTheme from '../theme/wallet-dark'
 import walletLightTheme from '../theme/wallet-light'
 import BraveCoreThemeProvider from '../../common/BraveCoreThemeProvider'
-import { ConnectWithSite } from '../components/extension'
-import { StyledExtensionWrapper } from '../stories/style'
 import store from './store'
 import * as WalletPanelActions from './actions/wallet_panel_actions'
-import { WalletState, PanelState, WalletPanelState, WalletAccountType } from '../constants/types'
-
-type Props = {
-  panel: PanelState
-  wallet: WalletState
-  actions: typeof WalletPanelActions
-}
-
-function mapStateToProps (state: WalletPanelState): Partial<Props> {
-  return {
-    panel: state.panel,
-    wallet: state.wallet
-  }
-}
-
-function mapDispatchToProps (dispatch: Dispatch): Partial<Props> {
-  return {
-    actions: bindActionCreators(WalletPanelActions, store.dispatch.bind(store))
-  }
-}
-
-const PanelWithState = connect(mapStateToProps, mapDispatchToProps)(Panel)
+import Container from './container'
 
 function App () {
   const [initialThemeType, setInitialThemeType] = React.useState<chrome.braveTheme.ThemeType>()
@@ -51,65 +27,10 @@ function App () {
         dark={walletDarkTheme}
         light={walletLightTheme}
       >
-        <PanelWithState />
+        <Container />
       </BraveCoreThemeProvider>
       }
     </Provider>
-  )
-}
-
-function Panel (props: Props) {
-  // TODO(petemill): If initial data or UI takes a noticeable amount of time to arrive
-  // consider rendering a "loading" indicator when `hasInitialized === false`, and
-  // also using `React.lazy` to put all the main UI in a separate JS bundle and display
-  // that loading indicator ASAP.
-  const [selectedAccounts, setSelectedAccounts] = React.useState<WalletAccountType[]>([
-    props.panel.accounts[0]
-  ])
-  const [readyToConnect, setReadyToConnect] = React.useState<boolean>(false)
-  const selectAccount = (account: WalletAccountType) => {
-    const newList = [...selectedAccounts, account]
-    setSelectedAccounts(newList)
-  }
-  const removeAccount = (account: WalletAccountType) => {
-    const newList = selectedAccounts.filter(
-      (accounts) => accounts.id !== account.id
-    )
-    setSelectedAccounts(newList)
-  }
-  const onSubmit = () => {
-    props.actions.connectToSite({
-      selectedAccounts,
-      siteToConnectTo: props.panel.connectedSiteOrigin
-    })
-  }
-  const primaryAction = () => {
-    if (!readyToConnect) {
-      setReadyToConnect(true)
-    } else {
-      onSubmit()
-    }
-  }
-  const secondaryAction = () => {
-    if (readyToConnect) {
-      setReadyToConnect(false)
-    } else {
-      props.actions.cancelConnectToSite()
-    }
-  }
-  return (
-    <StyledExtensionWrapper>
-      <ConnectWithSite
-        siteURL={props.panel.connectedSiteOrigin}
-        isReady={readyToConnect}
-        accounts={props.panel.accounts}
-        primaryAction={primaryAction}
-        secondaryAction={secondaryAction}
-        selectAccount={selectAccount}
-        removeAccount={removeAccount}
-        selectedAccounts={selectedAccounts}
-      />
-    </StyledExtensionWrapper>
   )
 }
